Feat: support erasure coding (#86)
* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Chore: make `cargo test` happy

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Feat: support erasure coding

* Chore: clean code

* Feat: support erasure coding

* Chore: add `Features` section

+++ helyim/src/operation/upload.rs
@@ -7,7 +7,7 @@ use crate::storage::Ttl;
 #[derive(Default, Serialize, Deserialize)]
 pub struct Upload {
     pub name: String,
-    pub size: u32,
+    pub size: usize,
     pub error: String,
 }
